Re-organize event structure
