+++ packages/core/src/__tests__/event/definitions.test.ts
@@ -11,7 +11,7 @@ import {
     DEFAULT_PRIORITY,
     ExtractEventData,
     EventNames,
-} from '../../event/event';
+} from '../../event/definition';
 
 interface TestUserEvents {
     'user:login': {
